Consolidate cpt-core tests
